refactor(MapMeasurementOverlay): break overlay and control widget apart
